# command-not-found: drop perl dependency