… nixos/modules/programs/command-not-found/command-not-found.pl