@Mic92 2019-12-01
I tested `export NIX_AUTO_INSTALL=1`, `export NIX_AUTO_RUN=1` and without any of those.