@bjornfor 2019-12-02
Small typo in commit subject.