@edolstra 2019-12-03
This to me is not really an improvement: it replaces a 51-line Perl script with a 141-line C++ program. As an aside, if we really want to get rid of Perl, then Rust is the way to go IMHO.

BTW, I think we can get rid of the auto-install feature, it was added as a bit of a joke and nobody uses it.