@Mic92 2019-12-03
> This to me is not really an improvement: it replaces a 51-line Perl script with a 141-line C++ program. As an aside, if we really want to get rid of Perl, then Rust is the way to go IMHO.
> 
> BTW, I think we can get rid of the auto-install feature, it was added as a bit of a joke and nobody uses it.

I choose C++: 
-  because it was proposed here: https://github.com/NixOS/nixpkgs/pull/48378#issuecomment-434900162 
-  it was accepted here: https://github.com/NixOS/nixpkgs/pull/68193
-  the sqlite interface is already in C, which made the building part quite forward and avoids having to cross-compile rust, when cross-compiling NixOS.

I am not sure the Rust version would be much shorter. The C++ is now 99 lines compared to 55 lines Perl. But it also generates more error messages.
I guess I just wait for `nix run` to become stable.