@Mic92 2019-12-06
Old closure size: 82.9M
New closure size: 33.6M

New one is faster but it probably hardly matters:

Old:
```
$ time command-not-found sqlite3
0.03s user 0.00s system 98% cpu 0.034 total
```

New:
```
$ time command-not-found sqlite3
0.00s user 0.00s system 87% cpu 0.004 total
```

The C++ program is now +28 lines longer.

Does this looks fine now?