@doronbehar 2020-05-21
BTW for the record, [`nix-index`](https://github.com/bennofs/nix-index) (Rust written) has a database and a "command-not-found" shell function which can be used:

https://github.com/bennofs/nix-index#usage-as-a-command-not-found-replacement

Though from my experience it feels somewhat slower to that of the original command-not-found implementation.