@Mic92 2020-05-21
There is also a rust version alternative: https://github.com/NixOS/nixpkgs/pull/88517